ProLogiX_analysis/requirements.txt
# requirements.txt

# Core dependencies
Flask
flask-cors

# Transformers and NLP
transformers
torch  
sentencepiece  # Required by BART and some other models
scipy  

# Utilities
gunicorn  

# Development
python-dotenv  
